Refactor Neovim class into 2 separate classes
* Neovim is now more similar to other ExtType classes (Buffer, Window,
Tabpage) - is now strictly neovim API wrappers
* NeovimClient handles session but extends Neovim so that we get an easy
interface
* Cleaned up some properties on Neovim Client
* Removed `apiPromise`, no longer needed for public interface.
BaseApi.request will now wait for a promise that resolves when API is
ready (e.g. after we request API info from neovim).
