+++ src/api/Window.test.js
@@ -29,9 +29,6 @@ describe('Window API', () => {
     );
 
     nvim = await attach({ proc });
-
-    await nvim.apiPromise;
-
     done();
   });
